+++ src/controllers/StudyMaterialController/material.controller.js
@@ -1,15 +1,16 @@
 import { notificationTypes } from "../../constants/notificationTypes.js";
 import { Course } from "../../models/CourseModel/course.model.js";
 import { Material } from "../../models/StudyMaterialModel/material.model.js";
-import { User } from "../../models/UserModel/user.model.js";
 import { notifyUsers } from "../../utils/NotificationEngine/notificationService.js";
 
 // Upload material
 
 export const uploadMaterial = async (req, res) => {
     try {
-        const id=req.params.id;
-        const { title, description, url, uploader } = req.body;
+        const id = req.params.id;
+        const { title, description, url } = req.body;
+
+        const faculty = req.dbUser;
 
         const course = await Course.findById(id);
 
@@ -19,12 +20,12 @@ export const uploadMaterial = async (req, res) => {
                 message: "Course not found"
             });
 
-        const user = await User.findById(uploader);
+        const isFaculty = course.faculties.some(t => t.toString() === faculty._id.toString());
 
-        if (!user)
-            return res.status(404).json({
+        if (!isFaculty)
+            return res.status(403).json({
                 success: false,
-                message: "User not found"
+                message: "You are not teaching this course"
             });
 
         if (!title || !description || !url)
@@ -33,28 +34,18 @@ export const uploadMaterial = async (req, res) => {
                 message: "Title, Description and URL required"
             });
 
-        const isTeacher = course.teachers.some(
-            t => t.toString() === uploader.toString()
-        );
-
-        if (!isTeacher)
-            return res.status(403).json({
-                success: false,
-                message: "You are not an instructor of this course"
-            });
-
-        const material=await Material.create({
+        const material = await Material.create({
             course: id,
             title,
             description,
             url,
-            uploader: uploader
+            uploader: faculty._id
         });
 
         if (course.students && course.students.length > 0) {
             await notifyUsers({
                 receivers: course.students,
-                sender: uploader,
+                sender: faculty._id,
                 type: notificationTypes.newStudyMaterial,
                 title: "New Study Material",
                 message: `${title} has been uploaded.`,
@@ -75,22 +66,32 @@ export const uploadMaterial = async (req, res) => {
 
 // Course wise materials
 
-export const courseMaterials=async(req,res)=>{
+export const courseMaterials = async (req, res) => {
     try {
-        const id=req.params.id;
-        
-        const course=await Course.findById(id);
+        const id = req.params.id;
 
-        if(!course)
+        const user = req.dbUser;
+
+        const course = await Course.findById(id);
+
+        if (!course)
             return res.status(404).json({
                 success: false,
                 message: "Course not found"
             });
-        
-        const materials=await Material.find({course:id})
+
+        const isOwner = (course.faculties.some(t => t.toString() === user._id.toString())) || (course.students.some(s => s.toString() === user._id.toString()));
+
+        if (!isOwner)
+            return res.status(403).json({
+                success: false,
+                message: "You are not authorized to get this course materials"
+            });
+
+        const materials = await Material.find({ course: id });
 
         return res.status(200).json({
-            success:true,
+            success: true,
             materials
         })
     } catch (error) {
@@ -100,26 +101,38 @@ export const courseMaterials=async(req,res)=>{
 
 // Material deletion
 
-export const deleteMaterial=async(req,res)=>{
+export const deleteMaterial = async (req, res) => {
     try {
-        const id=req.params.id;
+        const id = req.params.id;
 
-        const material=await Material.findById(id);
+        const faculty=req.dbUser;
 
-        if(!material)
+        const material = await Material.findById(id);
+
+        if (!material)
             return res.status(404).json({
-                success:false,
-                message:"Material not found"
+                success: false,
+                message: "Material not found"
             });
-        
-        await Material.deleteOne({_id:id});
+
+        const course=await Course.findById(material.course);
+
+        const isFaculty = course.faculties.some(t => t.toString() === faculty._id.toString());
+
+        if (!isFaculty)
+            return res.status(403).json({
+                success: false,
+                message: "You are not authorized to delete this course material"
+            });
+
+        await Material.deleteOne({ _id: id });
 
         return res.status(200).json({
-            success:true,
-            message:"Material deleted successfully"
+            success: true,
+            message: "Material deleted successfully"
         });
 
-   } catch (error) {
-        res.status(500).json({message:error.message});      
+    } catch (error) {
+        res.status(500).json({ message: error.message });
     }
 };
